RR-11415 Changes for the second cycle
FCRA query is using non-FCRA functonality, this needs to be changed.
Since this is a new query, removing it from the release.
Note: some of the attributes (Roylty.*, iesp.*, etc.) are left intact:
there's no harm to keep them in release branch; only code directly relevant
to the query is deleted.